RDRP-555: function moved & precommit tweaks
