+++ src/developer_config.yaml
@@ -25,7 +25,7 @@ global:
   output_long_form: False
   output_short_form: False
   output_gb_sas: True
-  output_tau: True
+  output_tau: False
   output_intram_by_pg: False
   output_intram_by_itl1: False
   output_intram_by_civil_defence: False
