+++ src/outputs/outputs_main.py
@@ -12,7 +12,6 @@ from src.outputs.intram_by_pg import output_intram_by_pg
 from src.outputs.intram_by_itl1 import output_intram_by_itl1
 from src.outputs.intram_by_civil_defence import output_intram_by_civil_defence
 from src.outputs.intram_by_sic import output_intram_by_sic
-from src.outputs.outputs_helpers import postcode_topup
 from src.outputs.total_fte import qa_output_total_fte
 
 OutputMainLogger = logging.getLogger(__name__)
@@ -83,14 +82,14 @@ def run_outputs(
         )
         OutputMainLogger.info("Finished status filtered output.")
 
-    outputs_df = outputs_df.astype({"postcodes_harmonised": "str"})
-    outputs_df["postcodes_harmonised"] = outputs_df["postcodes_harmonised"].apply(
-        postcode_topup
-    )
+    # outputs_df = outputs_df.astype({"postcodes_harmonised": "str"})
+    # outputs_df["postcodes_harmonised"] = outputs_df["postcodes_harmonised"].apply(
+    #     postcode_topup
+    # )
 
-    tau_outputs_df = tau_outputs_df.astype({"postcodes_harmonised": "str"})
-    tau_outputs_df["postcodes_harmonised"] = (
-        tau_outputs_df["postcodes_harmonised"].apply(postcode_topup))
+    # tau_outputs_df = tau_outputs_df.astype({"postcodes_harmonised": "str"})
+    # tau_outputs_df["postcodes_harmonised"] = (
+    #     tau_outputs_df["postcodes_harmonised"].apply(postcode_topup))
 
     # Running short form output
     if config["global"]["output_short_form"]:
